backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
gitpython>=3.1.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
httpx>=0.26.0
